requirements.txt
django-postgres-copy
djangorestframework
markdown
django-filter